Add helper functions to be implemented
Checking the recent solutions of the trainees they seemed to have issues with some of the exercises. Investigating these exercises they seemed to introduce multiple new concepts at the same time, occasionally making it harder for the trainee to solve them.

This PR tries to help this by adding a couple new exercises to the mix where the trainees need to solve a subproblem first, hopefully helping them realize how to solve a later problem. They could also re-use the functions in later exercises as well.